Moved viewModel to Models folder
