+++ Keas.Mvc/Models/TeamAdminMembersListModel.cs
@@ -1,0 +1,57 @@
+﻿using Keas.Core.Domain;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Keas.Mvc.Models
+{
+    public class TeamAdminMembersListModel
+    {
+
+        public Team Team { get; set; }
+
+        public List<UserRole> UserRoles { get; set; }
+
+        public static TeamAdminMembersListModel Create(Team team)
+        {
+            var viewModel = new TeamAdminMembersListModel()
+            {
+                Team = team,
+                UserRoles = new List<UserRole>()
+            };
+
+            foreach (var teamPermission in team.TeamPermissions)
+            {
+                if (viewModel.UserRoles.Any(a => a.User.Id == teamPermission.User.Id))
+                {
+                    viewModel.UserRoles.Single(a => a.User.Id == teamPermission.User.Id).Roles
+                        .Add(teamPermission.TeamRole);
+                }
+                else
+                {
+                    viewModel.UserRoles.Add(new UserRole(teamPermission));
+                }
+            }
+            return viewModel;
+        }
+
+    }
+
+    public class UserRole
+    {
+        public User User { get; set; }
+
+        public IList<Role> Roles { get; set; }
+
+        public UserRole(TeamPermission teamPermission)
+        {
+            User = teamPermission.User;
+            Roles = new List<Role>();
+            Roles.Add(teamPermission.TeamRole);
+        }
+
+        public string RolesList
+        {
+            get { return string.Join(", ", Roles.OrderBy(x => x.Name).Select(a => a.Name).ToArray()); }
+        }
+    }
+}
