+++ Keas.Mvc/Views/TeamAdmin/Members.cshtml
@@ -1,4 +1,4 @@
-﻿@model Keas.Mvc.Controllers.TeamAdminController.TeamAdminMembersListModel
+﻿@model Keas.Mvc.Models.TeamAdminMembersListModel
 
 @{
         ViewData["Title"] = "Members ";
